# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TSI API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/cidade": {
            "get": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Obter lista de cidades",
                "operationId": "ecba5880ce1bf48aeb9f4c2ced6e998b",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Lista de cidades retornada"
                                        },
                                        "cidades": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CidadeSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Criar uma nova cidade",
                "operationId": "3889c0e5504e87be63a884cf59822b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CidadeRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Cidade criada"
                                        },
                                        "cidade": {
                                            "$ref": "#/components/schemas/CidadeSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/cidade/{id}": {
            "get": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Obter os detalhes de uma cidade",
                "operationId": "ed6cc68c4016e65158046ef622df9c14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da cidade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CidadeSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Atualizar uma cidade",
                "operationId": "b6797ae33c8564e199d6708df99e4495",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da cidade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CidadeRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Cidade atualizada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cidades"
                ],
                "summary": "Deletar uma cidade",
                "operationId": "6a8758f55836e0b9e90475219c806667",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da cidade",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Registro não encontrado"
                    }
                }
            }
        },
...
        "/modalidade/{id}": {
            "get": {
                "tags": [
                    "Modalidades"
                ],
                "summary": "Detalhes da modalidade",
                "description": "Retorna os detalhes de uma modalidade.",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da modalidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModalidadeSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Modalidade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Modalidades"
                ],
                "summary": "Atualizar modalidade",
                "description": "Atualiza uma modalidade existente.",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da modalidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados da modalidade",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModalidadeRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModalidadeSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Modalidade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Modalidades"
                ],
                "summary": "Deletar modalidade",
                "description": "Deleta uma modalidade existente.",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da modalidade",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucesso"
                    },
                    "404": {
                        "description": "Modalidade não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CidadeRequestBody": {
                "title": "Corpo da Requisição para Cidade",
                "description": "Esquema de dados para o corpo da requisição ao criar ou atualizar uma cidade",
                "properties": {
                    "nome": {
                        "description": "Nome da cidade",
                        "type": "string",
                        "example": "São Paulo"
                    }
                },
                "type": "object"
            },
            "CidadeSchema": {
                "title": "Esquema de Cidade",
                "description": "Esquema de dados para o recurso de Cidade",
                "properties": {
                    "id": {
                        "description": "ID da cidade",
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome da cidade",
                        "type": "string",
                        "example": "São Paulo"
                    },
                    "estado": {
                        "$ref": "#/components/schemas/EstadoSchema"
                    }
                },
                "type": "object"
            },
...
            "EstadoSchema": {
                "title": "Esquema de Estado",
                "description": "Esquema de dados para o recurso de Estado",
                "properties": {
                    "id": {
                        "description": "ID do estado",
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do estado",
                        "type": "string",
                        "example": "São Paulo"
                    },
                    "uf": {
                        "description": "Sigla do estado",
                        "type": "string",
                        "example": "SP"
                    },
                    "pais": {
                        "$ref": "#/components/schemas/PaisSchema"
                    },
                    "cidades": {
                        "description": "Lista de cidades do estado",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CidadeSchema"
                        }
                    }
                },
                "type": "object"
            },
            "ModalidadeRequestBody": {
                "title": "Corpo da Requisição para Modalidade",
                "description": "Esquema de dados para o corpo da requisição ao criar ou atualizar uma modalidade",
                "properties": {
                    "nome": {
                        "description": "Nome da modalidade",
                        "type": "string",
                        "example": "Natação"
                    },
                    "descricao": {
                        "description": "Descrição da modalidade",
                        "type": "string",
                        "example": "Modalidade esportiva de natação"
                    },
                    "photo_path": {
                        "description": "Caminho da foto da modalidade",
                        "type": "string",
                        "example": "/path/to/photo.jpg"
                    }
                },
                "type": "object"
            },
            "ModalidadeSchema": {
                "title": "Modalidade",
                "description": "Esquema de dados para a modalidade",
                "properties": {
                    "id": {
                        "description": "ID da modalidade",
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome da modalidade",
                        "type": "string",
                        "example": "Natação"
                    },
                    "descricao": {
                        "description": "Descrição da modalidade",
                        "type": "string",
                        "example": "Modalidade esportiva de natação"
                    },
                    "photo_path": {
                        "description": "Caminho da foto da modalidade",
                        "type": "string",
                        "example": "/path/to/photo.jpg"
                    }
                },
                "type": "object"
            },
...
            "PaisSchema": {
                "title": "Esquema de País",
                "description": "Esquema de dados para o recurso de País",
                "properties": {
                    "identify": {
                        "description": "ID do país",
                        "type": "integer",
                        "example": "1"
                    },
                    "nome": {
                        "description": "Nome do país",
                        "type": "string",
                        "example": "Brasil"
                    },
                    "estado": {
                        "$ref": "#/components/schemas/EstadoSchema"
                    }
                },
                "type": "object"
            },
...
        }
    }
}